Single runHTMLTest
