+++ __tests__/html/offlineUI.fatalError.js
@@ -3,9 +3,6 @@
  */
 
 describe('offline UI', () => {
-  test('should show "Render error" connectivity status when a JavaScript error is present in the code.', async () => {
-    const { driver } = await loadHTMLTest('offlineUI.fatalError.html');
-
-    await expect(driver).resolves.toRunToCompletion({ ignoreConsoleError: true, ignorePageError: true });
-  });
+  test('should show "Render error" connectivity status when a JavaScript error is present in the code.', () =>
+    runHTMLTest('offlineUI.fatalError.html', { ignoreConsoleError: true }));
 });
