+++ __tests__/html/offlineUI.firstConnect.js
@@ -3,9 +3,6 @@
  */
 
 describe('offline UI', () => {
-  test('should display the "Connecting..." connectivity status when connecting for the first time', async () => {
-    const { driver } = await loadHTMLTest('offlineUI.firstConnect.html');
-
-    await expect(driver).resolves.toRunToCompletion();
-  });
+  test('should display the "Connecting..." connectivity status when connecting for the first time', () =>
+    runHTMLTest('offlineUI.firstConnect.html'));
 });
