+++ __tests__/html/offlineUI.networkInterrupt.js
@@ -3,9 +3,6 @@
  */
 
 describe('offline UI', () => {
-  test('should display "Network interruption occurred. Reconnecting…" status when connection is interrupted', async () => {
-    const { driver } = await loadHTMLTest('offlineUI.networkInterrupt.html');
-
-    await expect(driver).resolves.toRunToCompletion();
-  });
+  test('should display "Network interruption occurred. Reconnecting…" status when connection is interrupted', () =>
+    runHTMLTest('offlineUI.networkInterrupt.html'));
 });
