+++ __tests__/html/offlineUI.sendFailed.noAck.js
@@ -3,9 +3,6 @@
  */
 
 describe('offline UI', () => {
-  test('should display "Send failed. Retry" when activity is sent but not acknowledged', async () => {
-    const { driver } = await loadHTMLTest('offlineUI.sendFailed.noAck.html');
-
-    await expect(driver).resolves.toRunToCompletion();
-  });
+  test('should display "Send failed. Retry" when activity is sent but not acknowledged', () =>
+    runHTMLTest('offlineUI.sendFailed.noAck.html'));
 });
